+++ lib/torchvision/transforms/resize.rb
@@ -1,0 +1,13 @@
+module TorchVision
+  module Transforms
+    class Resize < Torch::NN::Module
+      def initialize(size)
+        @size = size
+      end
+
+      def forward(img)
+        F.resize(img, @size)
+      end
+    end
+  end
+end