+++ test/datasets_test.rb
@@ -4,7 +4,7 @@ class DatasetsTest < Minitest::Test
   def test_mnist
     trainset = TorchVision::Datasets::MNIST.new(root, train: true, download: true)
     assert_equal 60000, trainset.size
-    assert_kind_of Torch::Tensor, trainset[0][0]
+    assert_kind_of Vips::Image, trainset[0][0]
     assert_equal 5, trainset[0][1]
 
     assert_kind_of Torch::Tensor, trainset.data
@@ -14,14 +14,14 @@ class DatasetsTest < Minitest::Test
 
     testset = TorchVision::Datasets::MNIST.new(root, train: false, download: true)
     assert_equal 10000, testset.size
-    assert_kind_of Torch::Tensor, testset[0][0]
+    assert_kind_of Vips::Image, testset[0][0]
     assert_equal 7, testset[0][1]
   end
 
   def test_fashion_mnist
     trainset = TorchVision::Datasets::FashionMNIST.new(root, train: true, download: true)
     assert_equal 60000, trainset.size
-    assert_kind_of Torch::Tensor, trainset[0][0]
+    assert_kind_of Vips::Image, trainset[0][0]
     assert_equal 9, trainset[0][1]
 
     assert_kind_of Torch::Tensor, trainset.data
@@ -31,14 +31,14 @@ class DatasetsTest < Minitest::Test
 
     testset = TorchVision::Datasets::FashionMNIST.new(root, train: false, download: true)
     assert_equal 10000, testset.size
-    assert_kind_of Torch::Tensor, testset[0][0]
+    assert_kind_of Vips::Image, testset[0][0]
     assert_equal 9, testset[0][1]
   end
 
   def test_kmnist
     trainset = TorchVision::Datasets::KMNIST.new(root, train: true, download: true)
     assert_equal 60000, trainset.size
-    assert_kind_of Torch::Tensor, trainset[0][0]
+    assert_kind_of Vips::Image, trainset[0][0]
     assert_equal 8, trainset[0][1]
 
     assert_kind_of Torch::Tensor, trainset.data
@@ -48,7 +48,7 @@ class DatasetsTest < Minitest::Test
 
     testset = TorchVision::Datasets::KMNIST.new(root, train: false, download: true)
     assert_equal 10000, testset.size
-    assert_kind_of Torch::Tensor, testset[0][0]
+    assert_kind_of Vips::Image, testset[0][0]
     assert_equal 2, testset[0][1]
   end
 
@@ -57,6 +57,7 @@ class DatasetsTest < Minitest::Test
     assert_equal 50000, trainset.size
     assert_equal 6, trainset[0][1]
 
+    # returns Numo::NArray instead of Torch::Tensor like MNIST
     assert_kind_of Numo::NArray, trainset.data
     assert_equal [50000, 32, 32, 3], trainset.data.shape
     assert_equal [59, 62, 63], trainset.data[0..2].to_a